[refactor] move flask specific classes/functions etc in flask package
